+++ Config.xcconfig
@@ -1,0 +1,11 @@
+//
+//  Config.xcconfig
+//  Pillanner
+//
+//  Created by 윤규호 on 3/7/24.
+//
+
+// Configuration settings file format documentation can be found at:
+// https://help.apple.com/xcode/#/dev745c5c974
+
+KAKAO_NATIVE_APP_KEY = 60d9f983d2eb902984073b1401cb6a70
