+++ Pillanner/SignUp/LoginViewController.swift
@@ -14,7 +14,7 @@ import CryptoKit
 
 class LoginViewController: UIViewController, ASAuthorizationControllerDelegate, ASAuthorizationControllerPresentationContextProviding {
     
-    fileprivate var currentNonce: String?
+    var currentNonce: String?
     private let sidePaddingValue = 20
     private let paddingBetweenComponents = 30
 
@@ -77,12 +77,14 @@ class LoginViewController: UIViewController, ASAuthorizationControllerDelegate, 
     private let kakaoLoginButton: UIButton = {
         let button = UIButton()
         button.setImage(UIImage(named: "KakaoLoginImage"), for: .normal)
+        button.addTarget(target, action: #selector(kakaoLoginButtonTapped), for: .touchUpInside)
         return button
     }()
     
     private let appleLoginButton: UIButton = {
         let button = UIButton()
         button.setImage(UIImage(named: "AppleLoginImage"), for: .normal)
+        button.addTarget(target, action: #selector(appleLoginButtonTapped), for: .touchUpInside)
         return button
     }()
     
@@ -280,11 +282,10 @@ class LoginViewController: UIViewController, ASAuthorizationControllerDelegate, 
     
     @objc func signInButtonTapped() {
         print(#function)
-        // ????
-//        let signInVC = self.storyboard?.instantiateViewController(identifier: "SignUpViewController")
+
         let signUpVC = SignUpViewController()
         self.navigationController?.pushViewController(signUpVC, animated: true)
-//        present(SignUpViewController(), animated: true)
+
     }
     
     @objc func autoLoginButtonTapped(_ sender: UIButton) {
@@ -301,17 +302,6 @@ class LoginViewController: UIViewController, ASAuthorizationControllerDelegate, 
         }
     }
     
-    @objc func appleLoginButtonTapped() {
-        startSignInWithApple()
-    }
-    
-    @objc func naverLoginButtonTapped() {
-        
-    }
-    
-    @objc func kakaoLoginButtonTapped() {
-        
-    }
     
     @objc func findIDButtonTapped() {
         self.navigationController?.pushViewController(FindIDViewController(), animated: true)
@@ -319,98 +309,6 @@ class LoginViewController: UIViewController, ASAuthorizationControllerDelegate, 
     
     @objc func findPasswordButtonTapped() {
         self.navigationController?.pushViewController(FindPWViewController(), animated: true)
-    }
-    
-    // Apple Login 관련
-    private func randomNonceString(length: Int = 32) -> String {
-        precondition(length > 0)
-        var randomBytes = [UInt8](repeating: 0, count: length)
-        let errorCode = SecRandomCopyBytes(kSecRandomDefault, randomBytes.count, &randomBytes)
-        if errorCode != errSecSuccess {
-            fatalError(
-                "Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)"
-            )
-        }
-        
-        let charset: [Character] =
-        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-        
-        let nonce = randomBytes.map { byte in
-            // Pick a random character from the set, wrapping around if needed.
-            charset[Int(byte) % charset.count]
-        }
-        
-        return String(nonce)
-    }
-    
-    private func sha256(_ input: String) -> String {
-        let inputData = Data(input.utf8)
-        let hashedData = SHA256.hash(data: inputData)
-        let hashString = hashedData.compactMap {
-            return String(format: "%02x", $0)
-        }.joined()
-        
-        return hashString
-    }
-    
-    private func startSignInWithApple() {
-        let nonce = randomNonceString()
-        currentNonce = nonce
-        let appleIDProvider = ASAuthorizationAppleIDProvider()
-        let request = appleIDProvider.createRequest()
-        // request 요청을 했을 때 none가 포함되어서 릴레이 공격을 방지
-        // 추후 파베에서도 무결성 확인을 할 수 있게끔 함
-        request.requestedScopes = [.fullName, .email]
-        request.nonce = sha256(nonce)
-        
-        let authorizationController = ASAuthorizationController(authorizationRequests: [request])
-        authorizationController.delegate = self
-        authorizationController.presentationContextProvider = self
-        authorizationController.performRequests()
-    }
-    
-    
-    
-    func authorizationController(controller: ASAuthorizationController, didCompleteWithAuthorization authorization: ASAuthorization) {
-        if let appleIDCredential = authorization.credential as? ASAuthorizationAppleIDCredential {
-            guard let nonce = currentNonce else {
-                fatalError("Invalid state: A login callback was received, but no login request was sent.")
-            }
-            guard let appleIDToken = appleIDCredential.identityToken else {
-                print("Unable to fetch identity token")
-                return
-            }
-            guard let idTokenString = String(data: appleIDToken, encoding: .utf8) else {
-                print("Unable to serialize token string from data: \(appleIDToken.debugDescription)")
-                return
-            }
-            // Initialize a Firebase credential, including the user's full name.
-            let credential = OAuthProvider.appleCredential(withIDToken: idTokenString,
-                                                           rawNonce: nonce,
-                                                           fullName: appleIDCredential.fullName)
-            // Sign in with Firebase.
-            Auth.auth().signIn(with: credential) { (authResult, error) in
-                if let error = error {
-                    // Error. If error.code == .MissingOrInvalidNonce, make sure
-                    // you're sending the SHA256-hashed nonce as a hex string with
-                    // your request to Apple.
-                    print(error.localizedDescription)
-                    return
-                }
-                print("로그인에 성공했습니다.")
-                // User is signed in to Firebase with Apple.
-                // ...
-            }
-        }
-    }
-    
-    func authorizationController(controller: ASAuthorizationController, didCompleteWithError error: Error) {
-        // Handle error.
-        print("Sign in with Apple errored: \(error)")
-    }
-    
-    func presentationAnchor(for controller: ASAuthorizationController) -> ASPresentationAnchor {
-        return self.view.window!
     }
 }
